ci: Handle case where prebuilt was cached but brew install binutils
We create a simlink in prebuilt for certain tools like objcopy
that are brew installed from binutils.  It is likely that we
will cache the simlink, but still need to brew install binutils
this will fail because we will try to create the symlink again.
This change always recreates the symlink if the brew install was
required.

Signed-off-by: Brennan Ashton <bashton@brennanashton.com>